+++ components/list-group/src/overrides/item.js
@@ -5,7 +5,22 @@ import { Item as ListItem } from '@westpac/list';
 
 const Item = (props) => <ListItem {...props} />;
 
-const itemStyles = () => ({});
+const itemStyles = () => {
+	const { COLORS } = useBrand();
+
+	return {
+		label: 'listGroup-item',
+		margin: 0,
+		borderBottom: `1px solid ${COLORS.border}`,
+		display: 'flex',
+		alignItems: 'center',
+		justifyContent: 'space-between',
+
+		'@media print': {
+			borderColor: '#000',
+		},
+	};
+};
 
 const itemAttributes = () => null;