+++ components/selector/src/overrides/icon.js
@@ -22,6 +22,7 @@ const iconStyles = () => {
 		label: getLabel('selector-option-icon'),
 		marginRight: SPACING(4),
 		marginBottom: [SPACING(1), null, 0],
+		flex: 'none',
 	});
 };