+++ components/alert/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @westpac/alert
 
+## 1.2.1
+
+### Patch Changes
+
+- Updated dependencies [8058c8c3]
+- Updated dependencies [6a9ef63b]
+- Updated dependencies [0b22b4a0]
+- Updated dependencies [d27d24ca]
+  - @westpac/body@1.4.0
+  - @westpac/button@1.2.0
+  - @westpac/core@2.0.0
+  - @westpac/heading@1.0.1
+  - @westpac/icon@1.3.1
+
 ## 1.2.0
 
 ### Minor Changes