+++ components/form-pod/CHANGELOG.md
@@ -1,6 +1,15 @@
 # @westpac/form-pod
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [6a9ef63b]
+  - @westpac/core@2.0.0
+  - @westpac/icon@1.3.1
+
 ## 1.0.0-alpha.2
+
 ### Patch Changes
 
 - 520c2fc4: fix distribution files